+++ src/parlant/api/sessions.py
@@ -204,7 +204,6 @@ class SessionDTO(
 SessionCreationParamsCustomerIdField: TypeAlias = Annotated[
     CustomerId | None,
     Field(
-        default=None,
         description=" ID of the customer this session belongs to. If not provided, a guest customer will be created.",
         examples=[None, "cust_123xy"],
     ),